+++ hookee/plugins/request_query_params.py
@@ -1,11 +1,12 @@
-from hookee import pluginmanager
+from hookee.pluginmanager import REQUEST_PLUGIN
+from hookee.util import PrintUtil
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
 __version__ = "1.2.2"
 
-plugin_type = pluginmanager.REQUEST_PLUGIN
-print_util = None
+plugin_type = REQUEST_PLUGIN
+print_util = None  # type: PrintUtil
 
 
 def setup(hookee_manager):