+++ hookee/plugins/request_url_info.py
@@ -1,13 +1,14 @@
 from datetime import datetime
 
-from hookee import pluginmanager
+from hookee.pluginmanager import REQUEST_PLUGIN
+from hookee.util import PrintUtil
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
 __version__ = "1.2.2"
 
-plugin_type = pluginmanager.REQUEST_PLUGIN
-print_util = None
+plugin_type = REQUEST_PLUGIN
+print_util = None  # type: PrintUtil
 
 
 def setup(hookee_manager):
@@ -22,7 +23,8 @@ def run(request):
     timestamp = now.strftime("%m-%d-%Y %I:%M:%S %p")
 
     print_util.print_basic("[{}] \"{} {} {}\"".format(timestamp, request.method, request.base_url,
-                                                      request.environ["SERVER_PROTOCOL"]), color=print_util.request_color, bold=True)
+                                                      request.environ["SERVER_PROTOCOL"]),
+                           color=print_util.request_color, bold=True)
     print_util.print_basic()
 
     return request